requirements.txt
ibis-framework[duckdb]==8.0.0
pyarrow==14.0.1
pyyaml==6.0.1
pandas==2.1.4
python-dateutil==2.8.2
pandera==0.18.0

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1